Fixed to allow decimals

+++ BreakReminder/ViewModel/CountDownViewModel.cs
@@ -22,10 +22,10 @@ namespace BreakReminder.ViewModel
 
         #region Constructor
 
-        public CountDownViewModel(long numMin, NavigationViewModel navVM)
+        public CountDownViewModel(double numMin, NavigationViewModel navVM)
         {
             _timer = new DispatcherTimer();
-            _timerCounter = numMin * 60;
+            _timerCounter = (long) (numMin * 60);
             _time = new DateTime(2015, 1, 1, 0, 0, 0);
             _time = _time.AddSeconds(_timerCounter);
             _navVM = navVM;
